changing database

+++ migrations/20151216120019_migration_water.js
@@ -4,7 +4,7 @@ exports.up = function(knex, Promise) {
     table.increments('id').primary();
     table.integer('pi_id').references('pi.id');
 
-    table.integer('watering');
+    table.decimal('watering');
     table.decimal('available');
 
     table.timestamp('created_at').defaultTo(knex.fn.now());
